changing filters to /com/dowa/java/web/filters (still working in this practice); and renaming the Readme on HelloWorldWebApplication, which is done
