+++ practices/java-web/HelloWorldFilterWebApplication/.idea/libraries/Maven__org_apache_tomcat_tomcat_servlet_api_7_0_25.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.apache.tomcat:tomcat-servlet-api:7.0.25">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/tomcat/tomcat-servlet-api/7.0.25/tomcat-servlet-api-7.0.25.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/tomcat/tomcat-servlet-api/7.0.25/tomcat-servlet-api-7.0.25-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/tomcat/tomcat-servlet-api/7.0.25/tomcat-servlet-api-7.0.25-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
