+++ practices/java-web/HelloWorldFilterWebApplication/src/main/webapp/WEB-INF/web.xml
@@ -1,0 +1,32 @@
+<!DOCTYPE web-app PUBLIC
+        "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
+        "http://java.sun.com/dtd/web-app_2_3.dtd" >
+<web-app>
+    <filter>
+        <filter-name>filterOne</filter-name>
+        <filter-class>com.dowa.java.web.filters.FilterOne</filter-class>
+    </filter>
+    <filter-mapping>
+        <filter-name>filterOne</filter-name>
+        <url-pattern>/*</url-pattern>
+    </filter-mapping>
+
+    <filter>
+        <filter-name>filterTwo</filter-name>
+        <filter-class>com.dowa.java.web.filters.FilterTwo</filter-class>
+    </filter>
+    <filter-mapping>
+        <filter-name>filterTwo</filter-name>
+        <url-pattern>/</url-pattern>
+    </filter-mapping>
+
+    <servlet>
+        <servlet-name>myServlet</servlet-name>
+        <servlet-class>com.dowa.java.web.HelloWorldMvcServlet</servlet-class>
+        <load-on-startup>1</load-on-startup>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>myServlet</servlet-name>
+        <url-pattern>/hello-mvc-servlet</url-pattern>
+    </servlet-mapping>
+</web-app>
